// src/interfaces/interfaces.ts
export interface AssistanceFormValidators {
   valid: boolean;
   fio: {
      value: string;
      validators: {
         required: () => boolean;
      },
      valid: boolean;
      touched: boolean;
      errors: {
         required: boolean;
      },
      blur: () => boolean;
   },
   name: {
      value: string;
      validators: {
         required: () => boolean;
      },
      valid: boolean;
      touched: boolean;
      errors: {
         required: boolean;
      },
      blur: () => boolean;
   },
   surname: {
      value: string;
      validators: {
         required: () => boolean;
      },
      valid: boolean;
      touched: boolean;
      errors: {
         required: boolean;
      },
      blur: () => boolean;
   },
   patronymic: {
      value: string;
      validators: {
         required: () => boolean;
      },
      valid: boolean;
      touched: boolean;
      errors: {
         required: boolean;
      },
      blur: () => boolean;
   },
   phone: {
      value: string;
      validators: {
         required: () => boolean;
         isPhone: () => boolean;
      },
      valid: boolean;
      touched: boolean;
      errors: {
         required: boolean;
         isPhone: boolean;
      },
      blur: () => boolean;
   },
   birth: {
      value: string;
      validators: {
         required: () => boolean;
         isDDMMYYYY: () => boolean;
      },
      valid: boolean;
      touched: boolean;
      errors: {
         required: boolean;
         isDDMMYYYY: boolean;
      },
      blur: () => boolean;
   },
   addr: {
      value: string;
      validators: {
         required: () => boolean;
      },
      valid: boolean;
      touched: boolean;
      errors: {
         required: boolean;
      },
      blur: () => boolean;
   },
   district: {
      value: string;
      validators: {
         required: () => boolean;
      },
      valid: boolean;
      touched: boolean;
      errors: {
         required: boolean;
      },
      blur: () => boolean;
   },
   street: {
      value: string;
      validators: {
         required: () => boolean;
      },
      valid: boolean;
      touched: boolean;
      errors: {
         required: boolean;
      },
      blur: () => boolean;
   },
   house: {
      value: string;
      validators: {
         required: () => boolean;
      },
      valid: boolean;
      touched: boolean;
      errors: {
         required: boolean;
      },
      blur: () => boolean;
   },
   flat: {
      value: string;
      validators: {
         required: () => boolean;
      },
      valid: boolean;
      touched: boolean;
      errors: {
         required: boolean;
      },
      blur: () => boolean;
   },
   people_num: {
      value: number;
      validators: {
         required: () => boolean;
      },
      valid: boolean;
      touched: boolean;
      errors: {
         required: boolean;
      },
      blur: () => boolean;
   },
   people_fio: {
      value: string[];
      valid: boolean;
      touched: boolean;
      blur: () => boolean;
   },
   invalids: {
      value: string;
      valid: boolean;
      touched: boolean;
      blur: () => boolean;
   },
   children: {
      value: string;
      valid: boolean;
      touched: boolean;
      blur: () => boolean;
   },
   children_age: {
      value: string[];
      valid: boolean;
      touched: boolean;
      blur: () => boolean;
   },
   food: {
      value: string;
      valid: boolean;
      touched: boolean;
      blur: () => boolean;
   },
   water: {
      value: string;
      valid: boolean;
      touched: boolean;
      blur: () => boolean;
   },
   drugs: {
      value: string;
      valid: boolean;
      touched: boolean;
      blur: () => boolean;
   },
   products_detail: {
      value: string;
      valid: boolean;
      touched: boolean;
      blur: () => boolean;
   },
   gigien: {
      value: string;
      valid: boolean;
      touched: boolean;
      blur: () => boolean;
   },
   gigien_num: {
      value: string;
      valid: boolean;
      touched: boolean;
      blur: () => boolean;
   },
   pampers: {
      value: string;
      valid: boolean;
      touched: boolean;
      blur: () => boolean;
   },
   diet: {
      value: string;
      valid: boolean;
      touched: boolean;
      blur: () => boolean;
   },
   pers_data_agreement: {
      value: boolean;
      valid: boolean;
      touched: boolean;
      blur: () => boolean;
   },
   photo_agreement: {
      value: boolean;
      valid: boolean;
      touched: boolean;
      blur: () => boolean;
   },
}